+++ src/main/java/dev/willyelton/origins/client/renderer/CreatureRenderState.java
@@ -5,4 +5,5 @@ import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 
 public class CreatureRenderState extends LivingEntityRenderState {
     public EntityData entityData;
+    public boolean inCage;
 }