+++ src/main/java/dev/willyelton/origins/datagen/OriginsOfLifeLootTables.java
@@ -19,6 +19,7 @@ public class OriginsOfLifeLootTables extends VanillaBlockLoot {
         add(OriginsOfLife.FOSSIL_BLOCK_DEEPSLATE.get(), createOreDrop(OriginsOfLife.FOSSIL_BLOCK_DEEPSLATE.get(), OriginsOfLife.FOSSIL.get()));
         add(OriginsOfLife.FOSSIL_BLOCK_SULFUR.get(), createOreDrop(OriginsOfLife.FOSSIL_BLOCK_SULFUR.get(), OriginsOfLife.FOSSIL.get()));
         add(OriginsOfLife.FOSSIL_BLOCK_CLAY.get(), createOreDrop(OriginsOfLife.FOSSIL_BLOCK_CLAY.get(), OriginsOfLife.FOSSIL.get()));
+        dropSelf(OriginsOfLife.DISPLAY_CASE.get());
     }
 
     @Override